doc/doxygen.sh: add --ase and --quiet options

+++ doc/doxygen.sh
@@ -6,7 +6,16 @@ SCRIPTPATH="$(readlink -f "$0")" && SCRIPTDIR=${SCRIPTPATH%/*}
 # == Options, Setup ==
 BUILDDIR=$(readlink -f "${BUILDDIR:-out}")
 DOXYDIR="$BUILDDIR"/doxygen/
-mkdir -p $DOXYDIR/doxy/ && rm -rf $DOXYDIR/*/ # leaves parent dir + files
+QUIET=false
+WITH_ASE=false
+while test $# -ne 0 ; do
+  case "$1" in
+    --quiet)		QUIET=true ;;
+    --ase)		WITH_ASE=true ;;
+    *)                  die "unknown argument: '$1'" ;;
+  esac
+  shift
+done
 
 # == Tagfile ==
 ( cd $DOXYDIR
@@ -155,13 +164,16 @@ run_doxygen() # run_doxygen NAME NUMBER BRIEF OUTDIR
 )
 
 # == Ase Docs ==
-rm -rf $DOXYDIR/doxy && mkdir -p $DOXYDIR/doxy
-cp -a ase devices jsonipc $DOXYDIR/doxy/
-ASE_VERSION=anklang-$(git describe)
-ASE_NAME="${ASE_VERSION%-v*}" && ASE_NAME="${ASE_NAME^}"
-ASE_BRIEF="ASE — Anklang Sound Engine (C++)"
-run_doxygen +dot +ref +cc "$ASE_NAME" "${ASE_VERSION#*-v}" "$ASE_BRIEF" $DOXYDIR/ase/
-echo "TAGFILES += \"$DOXYDIR/ase/tagfile.xml=../ase/\"" >> $DOXYDIR/Doxyfile
+if $WITH_ASE ; then
+  rm -rf $DOXYDIR/doxy && mkdir -p $DOXYDIR/doxy
+  cp -a ase devices jsonipc $DOXYDIR/doxy/
+  ASE_VERSION=anklang-$(git describe)
+  ASE_NAME="${ASE_VERSION%-v*}" && ASE_NAME="${ASE_NAME^}"
+  ASE_BRIEF="ASE — Anklang Sound Engine (C++)"
+  run_doxygen +dot +ref +cc "$ASE_NAME" "${ASE_VERSION#*-v}" "$ASE_BRIEF" $DOXYDIR/ase/
+fi
+test -r "$DOXYDIR/ase/tagfile.xml" &&
+  echo "TAGFILES += \"$DOXYDIR/ase/tagfile.xml=../ase/\"" >> $DOXYDIR/Doxyfile
 
 # == Cleanup ==
 rm -rf $DOXYDIR/doxy
